d2::align::_align_element(): Don't cast log2(perturb) to unsigned int.

+++ d2/align.h
@@ -2331,7 +2331,7 @@ public:
 				perturb *= 0.5;
 
 				if (lod > 0 
-				 && lod > (unsigned int) lrint(log(perturb) / log(2)) - lod_preferred) {
+				 && lod > lrint(log(perturb) / log(2)) - lod_preferred) {
 
 					/*
 					 * Work with images twice as large
